Course Name Change...
Partly to get the correct prefixes and reflect the unknown course number, but also that I realized the title works much better the other way around. It adds a double interpretation that I like: we are either learning computer science through the lens of biological applications, or computing our way through the study of biology.